added submissions component
