+++ pages/contests/[id].tsx
@@ -3,6 +3,7 @@ import { useRouter } from 'next/router'
 import { useQuery } from '@apollo/client'
 import { GET_CONTEST, GET_CONTEST_SUBMISSIONS } from '../../lib/graphql'
 import { useState, useEffect } from 'react'
+import Submissions from '../../components/Submissions'
 
 export default function ContestDetailPage() {
   const router = useRouter()
@@ -30,10 +31,6 @@ export default function ContestDetailPage() {
   const submissionList = submissionData?.contest_submissions
 
   const [selectedImage, setSelectedImage] = useState(null)
-
-  const handleImageClick = (image) => {
-    setSelectedImage(image)
-  }
 
   const closeImageModal = () => {
     setSelectedImage(null)
@@ -84,29 +81,10 @@ export default function ContestDetailPage() {
                   {!loadingSubmission && !errorSubmission && (
                     <>
                       {submissionList?.map((image) => (
-                        <div
-                          key={image.id}
-                          className='w-1/4 mt-2 flex flex-col items-center mx-2'
-                        >
-                          <img
-                            src={image.picture.picture_path}
-                            alt={`Image ${image.id}`}
-                            className='w-full h-auto max-h-60'
-                            onClick={() => handleImageClick(image)}
-                          />
-                          <p className='mt-2 text-center'>
-                            User: {image.picture.user.name_first}{' '}
-                            {image.picture.user.name_last}
-                          </p>
-                          <div className=' flex items-center justify-center'>
-                            <button
-                              className='mt-2 text-gray-700 bg-gray-500 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2'
-                              type='submit'
-                            >
-                              Vote
-                            </button>
-                          </div>
-                        </div>
+                        <Submissions
+                          image={image}
+                          setSelectedImage={setSelectedImage}
+                        />
                       ))}
                     </>
                   )}
